restructure file folders
