+++ src/utils/plant-calculations.ts
